Move postgres integration to core, deprecate @appsignal/pg
