+++ packages/nodejs/src/instrumentation/pg/patches/index.ts
@@ -1,6 +1,7 @@
 import shimmer from "shimmer"
 import { Submittable } from "pg"
-import { Tracer, Span } from "@appsignal/nodejs"
+import { Tracer } from "../../../interfaces/tracer"
+import { Span } from "../../../interfaces/span"
 
 interface Handlers {
   handleError?: (msg: any) => any
